requirements.txt
aio-pika==9.4.2
aiormq==6.8.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
black==24.4.2
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
firebase-admin==6.5.0
firestore==0.0.8
flake8==7.1.0
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
iso8601==2.1.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
motor==3.5.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
odmantic==1.0.2
orjson==3.10.6
packaging==24.1
pamqp==3.3.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.7.0
pydantic-extra-types==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.18.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.8.0
pyparsing==3.1.2
pytest==8.2.2
pytest-asyncio==0.23.7
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
stripe==10.1.0
tenacity==8.4.2
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
